Refactor authentication and health check controllers; add AuthenticationDTO
