+++ src/main/java/com/adv_investor/auth_module/controllers/heath/HeathCheckController.java
@@ -1,4 +1,4 @@
-package com.adv_investor.auth_module.controllers;
+package com.adv_investor.auth_module.controllers.heath;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
