+++ src/main/java/com/adv_investor/auth_module/controllers/login/AuthenticationController.java
@@ -1,0 +1,27 @@
+package com.adv_investor.auth_module.controllers.login;
+
+import jakarta.validation.Valid;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+@RequestMapping("/auth")
+public class AuthenticationController {
+    @Autowired
+    private AuthenticationManager authenticationManager;
+
+    @PostMapping("/login")
+    public ResponseEntity login(@RequestBody @Valid AuthenticationDTO data){
+        var userPassword = new UsernamePasswordAuthenticationToken(data.login(), data.password());
+        var authentication = this.authenticationManager.authenticate(userPassword);
+
+        if (authentication == null) {
+            return ResponseEntity.badRequest().build();
+        }
+
+        return ResponseEntity.ok().build();
+    }
+}
