+++ src/main/java/com/adv_investor/auth_module/controllers/login/AuthenticationDTO.java
@@ -1,4 +1,4 @@
-package com.adv_investor.auth_module.domain.user;
+package com.adv_investor.auth_module.controllers.login;
 
 public record AuthenticationDTO(String login, String password) {
 }
