--- src/main/java/com/adv_investor/auth_module/domain/user/RegisterDTO.java
@@ -1,7 +1,0 @@
-package com.adv_investor.auth_module.domain.user;
-
-public record RegisterDTO(String cpfDocument,
-                          String name,
-                          String email,
-                          Boolean isActive,
-                          UserProfile profile) { }
