+++ src/main/java/com/adv_investor/auth_module/domain/user/User.java
@@ -7,10 +7,8 @@ import lombok.NoArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
-
 import java.util.Collection;
 import java.util.List;
-import java.util.Stack;
 
 @Table(name = "users")
 @Entity(name = "users")
