+++ src/main/java/com/adv_investor/auth_module/repositories/UserRepository.java
@@ -4,6 +4,5 @@ import com.adv_investor.auth_module.domain.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, String> {
-
     User findByCpfDocument(String cpfDocument);
 }
